feat: autorType 추가
