+++ lib/types/authorType.ts
@@ -1,0 +1,6 @@
+export type Author = {
+  username: string;
+  bio: string | null;
+  image: string;
+  following: boolean;
+};
