+++ lib/types/commentType.ts
@@ -1,3 +1,5 @@
+import { Author } from './authorType';
+
 export interface Comments {
   comments: CommentType[];
 }
@@ -9,11 +11,4 @@ export type CommentType = {
   slug: string;
   author: Author;
   updatedAt: number;
-};
-
-export type Author = {
-  username: string;
-  bio: string;
-  image: string;
-  following: boolean;
 };
